Appends 'antlr4' to the javaSource directory for generated Java code.
This fixes a problem that manifests itself when projects using this plugin are converted to IntelliJ projects using gen-idea. SBT expects plugins to put their generated source in a sub-directory of sourceManaged, named for the plugin. In order to support this, gen-idea appears to add the first level of directories under src_managed/main as "source paths" in the generated project, and so if you specify antlr4PackageName, the top level of the package hierarchy gets added as a project source path by accident. You can probably guess how well this works out.

Putting sbt-antlr4's generated code one level down, under 'antlr4', prevents this problem.